requirements.txt
pytest
memory-profiler
pillow
argparse

